refactor(backend): remove extra functions
